% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_intro.R
\docType{package}
\name{NNbenchmark-package}
\alias{NNbenchmark-package}
\alias{NNbenchmark}
\title{Package NNbenchmark}
\description{
Datasets and functions to benchmark (convergence, speed, ease of use) R packages 
dedicated to regression with neural networks (no classification in this version).
The templates for the tested packages are available at 
<https://github.com/pkR-pkR/NNbenchmarkTemplates/tree/master/R> and 
<https://akshajverma.com/NNbenchmarkWeb>.
}
\examples{
ds <- grep("^[m,u]", ls("package:NNbenchmark"), value = TRUE); ds
t(sapply(ds, function(x) dim(get(x)))) 
NNdataSummary(NNdatasets)
plot(uGauss2)
pairs(mIshigami)
ht(NNdatasets, n = 2, l = 6)
}
