% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNSummary.R
\name{NNSummary}
\alias{NNSummary}
\title{Summarize Calculations of RMSE, MSE, MAE, and WAE}
\usage{
NNSummary(y_pred, y0, dgts = 4, time)
}
\arguments{
\item{y_pred}{numeric vector of the predicted values}

\item{y0}{numeric vector of the observed values}

\item{dgts}{integer value for how many digits to round to}

\item{time}{numeric value of time}
}
\value{
A vector of RMSE, MSE, MAE, WAE, and time values for a single training.
}
\description{
Summarize measures of fit and time for a single training. Measures of fit 
include the Root Mean Squared Error (RMSE), the Mean Squared Error (MSE), 
the Mean Absolute Error (MAE), and the Worst Absolute Error (WAE) rounded
by default to 4 digits and set to \code{na.rm = TRUE}. See more at \code{\link{funRMSE}}. 
The summary can also include the results of time from \code{\link{getTimer}} in
NNbenchmark or results of time from other functions/packages
}
\examples{
options("digits.secs" = 4)
timeTT <- createTimer()

timeTT$start("event")
y0 <- 1:19
y_pred <- y0 + rnorm(length(y0), sd = 0.3)
timeTT$stop("event") 

time <- getTimer(timeTT)

NNSummary(y_pred, y0, 4, time[,4])

}
