% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTrain_Predict.R
\name{NNTrain_Predict}
\alias{NNTrain_Predict}
\alias{train_predict_1mth_1data}
\alias{train_predict_1data}
\title{Generic Functions for Training and Predicting}
\usage{
train_predict_1mth_1data(
  dset,
  method,
  trainFUN,
  hyperparamFUN,
  predictFUN,
  summaryFUN,
  prepareZZ.arg = list(),
  nrep = 5,
  doplot = FALSE,
  plot.arg = list(col1 = 1:nrep, lwd1 = 1, col2 = 4, lwd2 = 3),
  pkgname,
  pkgfun,
  rdafile = FALSE,
  odir = "~/",
  echo = FALSE,
  echoreport = 1,
  ...
)

train_predict_1data(
  dset,
  methodlist,
  trainFUN,
  hyperparamFUN,
  predictFUN,
  summaryFUN,
  closeFUN,
  startNN = NA,
  prepareZZ.arg = list(),
  nrep = 5,
  doplot = FALSE,
  plot.arg = list(),
  pkgname = "pkg",
  pkgfun = "train",
  rdafile = FALSE,
  odir = "~/",
  echo = FALSE,
  ...
)
}
\arguments{
\item{dset}{a number or string indicating which dataset to use, see \code{\link{NNdataSummary}}}

\item{method}{a method for a particular function}

\item{trainFUN}{the training function used}

\item{hyperparamFUN}{the function resulting in parameters needed for training}

\item{predictFUN}{the prediction function used}

\item{summaryFUN}{measure performance by observed and predicted y values, \code{\link{NNSummary}} is ready to use}

\item{prepareZZ.arg}{list of arguments for \code{\link{prepareZZ}}}

\item{nrep}{a number for how many times a neural network should be trained}

\item{doplot}{logical value, TRUE executes plots and FALSE does not}

\item{plot.arg}{list of arguments for plots}

\item{pkgname}{package name}

\item{pkgfun}{package function to train neural network}

\item{rdafile}{logical value,}

\item{odir}{directory}

\item{echo}{logical value, print summary if TRUE}

\item{echoreport}{separates training between packages with some text}

\item{...}{additional arguments}

\item{methodlist}{list of methods per package/function}

\item{closeFUN}{a function to detach packages or other necessary environment clearing}

\item{startNN}{a function to start needed outside libraries, for example, h2o}
}
\value{
A vector as in NNSummary for train_predict_1mth_1data and an array for train_predict_1data
}
\description{
An implementation of \code{do.call} so any neural network function that fits 
the format can be tested. 

In \code{train_predict_1mth_1data}, a neural network is trained on one dataset
and then used for predictions, with several functionalities. Then, the performance 
of the neural network is summarized.

\code{train_predict_1data} serves a wrapper function for \code{train_predict_1mth_1data} 
for multiple methods.
}
