% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{uGauss2}
\alias{uGauss2}
\title{Dataset uGauss2}
\description{
An univariate dataset (x, y) of class data.frame and dim 250 x 2 to be fitted by a 
neural network with 4 hidden neurons (13 parameters).
}
\examples{
ht(uGauss2)
plot(uGauss2)
}
\references{
Rust, B., NIST (1996) \url{https://itl.nist.gov/div898/strd/nls/data/gauss2.shtml}
}
\keyword{datasets}
